EC-165 - OPENMRS-EIP-DOCKER: Create a custom docker image for offline installation

+++ release_stuff/etc/eptssync/scripts/eptssync_inconsistence_check.sh
@@ -20,24 +20,28 @@ SHARED_DIR="$EIP_HOME/shared"
 RELEASES_PACKAGES_DIR="$SHARED_DIR/releases"
 JAR_FILE_NAME="eptssync-api-1.0-SNAPSHOT.jar"
 
-. $EIP_SETUP_SCRIPTS_DIR/release_info.sh
-echo "Verifying $RELEASE_NAME packages download status"
-$EIP_SETUP_SCRIPTS_DIR/download_release.sh "$RELEASES_PACKAGES_DIR" "$RELEASE_NAME" "$OPENMRS_EIP_APP_RELEASE_URL" "$EPTSSYNC_API_RELEASE_URL"
-
-CURRENT_RELEASES_PACKAGES_DIR="$RELEASES_PACKAGES_DIR/$RELEASE_NAME"
-
-RELEASE_PACKAGES_DOWNLOAD_COMPLETED="$CURRENT_RELEASES_PACKAGES_DIR/download_completed"
-if [ ! -f "$RELEASE_PACKAGES_DOWNLOAD_COMPLETED" ]
+if [ ! -f "$EPTSSYNC_HOME/$JAR_FILE_NAME" ]
 then
-    echo "Error trying to download release packages: $RELEASE_NAME. See previous messages."
-    echo "EPTS startup process failed"
-    exit 1
+    . $EIP_SETUP_SCRIPTS_DIR/release_info.sh
+    echo "Verifying $RELEASE_NAME packages download status"
+    $EIP_SETUP_SCRIPTS_DIR/download_release.sh "$RELEASES_PACKAGES_DIR" "$RELEASE_NAME" "$OPENMRS_EIP_APP_RELEASE_URL" "$EPTSSYNC_API_RELEASE_URL"
+
+    CURRENT_RELEASES_PACKAGES_DIR="$RELEASES_PACKAGES_DIR/$RELEASE_NAME"
+
+    RELEASE_PACKAGES_DOWNLOAD_COMPLETED="$CURRENT_RELEASES_PACKAGES_DIR/download_completed"
+    if [ ! -f "$RELEASE_PACKAGES_DOWNLOAD_COMPLETED" ]
+    then
+        echo "Error trying to download release packages: $RELEASE_NAME. See previous messages."
+        echo "EPTS startup process failed"
+        exit 1
+    fi
+
+    # copying files
+    EPTSSYNC_PACKAGE_RELEASE_FILE_NAME=$(echo "$EPTSSYNC_API_RELEASE_URL" | rev | cut -d'/' -f 1 | rev)
+    echo "Copying $EPTSSYNC_PACKAGE_RELEASE_FILE_NAME to $EPTSSYNC_HOME/$JAR_FILE_NAME"
+    cp "$CURRENT_RELEASES_PACKAGES_DIR/$EPTSSYNC_PACKAGE_RELEASE_FILE_NAME" "$EPTSSYNC_HOME/$JAR_FILE_NAME"
 fi
 
-# copying files
-EPTSSYNC_PACKAGE_RELEASE_FILE_NAME=$(echo "$EPTSSYNC_API_RELEASE_URL" | rev | cut -d'/' -f 1 | rev)
-echo "Copying $EPTSSYNC_PACKAGE_RELEASE_FILE_NAME to $EPTSSYNC_HOME/$JAR_FILE_NAME"
-cp "$CURRENT_RELEASES_PACKAGES_DIR/$EPTSSYNC_PACKAGE_RELEASE_FILE_NAME" "$EPTSSYNC_HOME/$JAR_FILE_NAME"
 
 cp $CONFIG_FILE_ORIGINAL $CONFIG_FILE
 
@@ -47,7 +51,6 @@ sed -i "s/openmrs_db_host/$openmrs_db_host/g" $CONFIG_FILE
 sed -i "s/openmrs_db_port/$openmrs_db_port/g" $CONFIG_FILE 
 sed -i "s/openmrs_db_name/$openmrs_db_name/g" $CONFIG_FILE 
 sed -i "s/spring_openmrs_datasource_password/$spring_openmrs_datasource_password/g" $CONFIG_FILE 
-
 
 # Start application.
 echo "Starting EPTS Application"
